requirements.txt
fastapi>=0.95.0
fastapi-limiter
slowapi>=0.1.9
uvicorn[standard]>=0.20.0
asyncpg>=0.27.0
alembic>=1.10.0
SQLAlchemy>=2.0.0
pydantic>=2.7.0,<3
pydantic-settings>=2.0.0
alembic>=1.12.0
email-validator>=1.3.0
jinja2>=3.1
python-multipart>=0.0.6
aiofiles>=23.1.0
aiohttp>=3.8.4
redis>=4.2
cloudinary>=1.30.0
python-dotenv>=1.0.0
bcrypt==4.0.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]>=3.3.0
psycopg2-binary